err/requirements.txt
Flask
google-generativeai
pdfplumber
